Feature/crud validators

+++ src/main/kotlin/domain/usecase/crud/route/DeleteRouteUseCase.kt
@@ -1,11 +1,25 @@
 package domain.usecase.crud.route
 
 import domain.repository.RouteRepository
+import domain.exception.DatabaseOperationFailedException
+import domain.validator.ValidationResult
+import domain.validator.routes.RouteIdValidator
 
 class DeleteRouteUseCase(
-    private val routeRepository: RouteRepository
+    private val routeRepository: RouteRepository,
+    private val idValidator: RouteIdValidator
 ) {
-    suspend operator fun invoke(id: String): Boolean {
-        return routeRepository.delete(id)
+    suspend operator fun invoke(id: String): ValidationResult<Unit> {
+        val validationResult = idValidator.validate(id)
+        if (validationResult is ValidationResult.Failure) {
+            return validationResult
+        }
+
+        val isDeleted = routeRepository.delete(id)
+        return if (isDeleted) {
+            ValidationResult.Success(Unit)
+        } else {
+            ValidationResult.Failure(listOf(DatabaseOperationFailedException("delete", "route")))
+        }
     }
 }
